requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
macholib==1.16.3
packaging==25.0
pillow==11.2.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
setuptools==80.9.0
